pyproject.toml
[tool.poetry]
name = "headhunter-bd"
version = "0.1.0"
description = ""
authors = ["Andrey Saksonov <asaksonov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
